pyproject.toml
[project]
name = "openai-app-example"
version = "0.1.0"
description = "Simple Typer CLI that sends a prompt to OpenAI and prints the response"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.7",
  "python-dotenv>=1.0",
  "openai>=1.0.0",
  "langfuse>=3.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
